pyproject.toml
[project]
name = "project-1-document-portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "omegaconf>=2.3.0",
    "pandas>=3.0.1",
    "pymupdf>=1.27.2",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "black>=26.3.0",
    "ipykernel>=7.2.0",
    "isort>=8.0.1",
]
